# perf: close sessions async