--- a/google-cloud-spanner/src/main/java/com/google/cloud/spanner/SessionPool.java
+++ b/google-cloud-spanner/src/main/java/com/google/cloud/spanner/SessionPool.java
@@ -1594,7 +1594,7 @@ final class SessionPool {
             }
           }
         },
-        MoreExecutors.directExecutor());
+        executor);
     return res;
   }