@olavloite 2020-01-09
The binary compatibility check fails because of:

```
[ERROR] 7012: com.google.cloud.spanner.Session: Method 'public com.google.api.core.ApiFuture asyncClose()' has been added to an interface
[ERROR] 7012: com.google.cloud.spanner.spi.v1.SpannerRpc: Method 'public com.google.api.core.ApiFuture asyncDeleteSession(java.lang.String, java.util.Map)' has been added to an interface
```

Both `Session` and `SpannerRpc` are public interfaces, but are only used internally in the client library. There are no public methods returning either a `Session` or a `SpannerRpc`.