@olavloite 2020-01-19
@skuruppu @hengfengli @chingor13 
This PR cannot be merged because the Binary Compatibility test fails. This happens because it adds a method to two public interfaces. These interfaces are not really public (as in: returned to users), but must be public because they reside in different packages within the Spanner client library and are used by different parts of the Spanner client lib.

So:
1. Should we still treat this as a breaking change?
1. Regardless of whether we do treat is a breaking change or not: What action is needed to make the PR mergeable?