@skuruppu 2020-01-23
> > > @skuruppu @hengfengli @chingor13
> > > This PR cannot be merged because the Binary Compatibility test fails. This happens because it adds a method to two public interfaces. These interfaces are not really public (as in: returned to users), but must be public because they reside in different packages within the Spanner client library and are used by different parts of the Spanner client lib.
> > > So:
> > > 
> > > 1. Should we still treat this as a breaking change?
> > 
> > 
> > I don't believe it should be flagged as a breaking change given that we should reserve those when making backwards-incompatible changes that requires users to rebuild their apps.
> > > 1. Regardless of whether we do treat is a breaking change or not: What action is needed to make the PR mergeable?
> > 
> > 
> > I'm not sure actually :( @kolea2 would you be able to advice us on whether it's ok to ignore a Binary Compatibility failure and merge this change.
> > I don't know why the Java 11 test fails so I triggered a rebuild.
> 
> As discussed offline, it's ok to push this change despite the Binary Compatibility failure since the changes being made are to internal interfaces.

Yikes, it turns out the compatibility check just keeps failing for all future PRs as well. I was hoping that  it would stop complaining after it was merged in. Let me check with @kolea2 offline about what we can do to fix this issue.