@skuruppu 2020-01-23
> Should we revert this change? We can wait until it is resolved so that we can merge it in.

Yeah good point. I'll revert it now.